+++ src/core/boolean.ts
@@ -1,7 +1,0 @@
-import type { RVBoolean } from "@/types/boolean"
-
-export function boolean(): RVBoolean {
-  return {
-    type: "boolean",
-  } as never
-}