+++ src/core/date.ts
@@ -1,7 +1,0 @@
-import type { RVDate } from "@/types/date"
-
-export function date(): RVDate {
-  return {
-    type: "date",
-  } as never
-}