+++ src/resolvers/typebox/date.ts
@@ -1,6 +1,0 @@
-import type { TDate } from "@sinclair/typebox";
-import { Type } from "@sinclair/typebox";
-
-export function TBDate(): TDate {
-	return Type.Date();
-}