+++ src/resolvers/typebox/object.ts
@@ -1,5 +1,0 @@
-import { type TObject, Type } from "@sinclair/typebox"
-
-export function TBObject(schema: TObject): TObject {
-  return Type.Object(schema)
-}