+++ src/resolvers/zod/array.ts
@@ -1,5 +1,0 @@
-import z, { type ZodArray } from "zod"
-
-export function ZDArray(schema: ZodArray): ZodArray {
-  return z.array(schema)
-}