+++ src/resolvers/zod/boolean.ts
@@ -1,5 +1,0 @@
-import z, { type ZodBoolean } from "zod"
-
-export function ZDBoolean(): ZodBoolean {
-  return z.boolean()
-}