+++ src/resolvers/zod/tuple.ts
@@ -1,7 +1,0 @@
-import z, { type ZodType } from "zod"
-
-export function ZDTuple(
-  schemas: [z.core.SomeType, ...z.core.SomeType[]],
-): ZodType {
-  return z.tuple(schemas)
-}