+++ src/types/date.d.ts
@@ -1,6 +1,0 @@
-import type { RVSchema } from './schema';
-
-export interface RVDate extends RVSchema {
-	type: 'date';
-	infer: Date;
-}