+++ src/types/schema.d.ts
@@ -1,6 +1,0 @@
-export interface RVSchema {
-  type: unknown
-  infer: unknown
-  schema: unknown
-  optional?: boolean
-}