+++ tsup.config.ts
@@ -1,11 +1,11 @@
-import { defineConfig } from "tsup"
+import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src", "!./src/**/*.spec.ts"],
-  clean: true,
+  entry: ["!./src/**/*.spec.ts"],
   target: "es2022",
   format: ["cjs", "esm"],
   splitting: false,
   bundle: true,
+  dts: true,
   external: ["@sinclair/typebox", "zod"],
-})
+});